+++ CHANGELOG.md
@@ -5,5 +5,7 @@
 ### Added
 
 - `dank build` and `dank serve` commands delegate to esbuild for website development
+- `dank.config.ts` configures api/web service commands that are started with `dank serve`
+  using the `DankConfig['services']` field of `defineConfig`
 
 [Unreleased]: https://github.com/eighty4/dank/commits/main